+++ usecase/src/main/kotlin/dev/agner/portfolio/usecase/index/model/IndexValue.kt
@@ -1,8 +1,9 @@
 package dev.agner.portfolio.usecase.index.model
 
 import kotlinx.datetime.LocalDate
+import java.math.BigDecimal
 
 data class IndexValue(
     val date: LocalDate,
-    val value: Double,
+    val value: BigDecimal,
 )
